# swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Pages API
  description: >
    ## Summary

    The Pages API is domain service for Pages

    ## Uploading an Asset

    1. A client requests to upload an asset to  `/documents/{documentId}/assets/request-upload` enpoint and it returns urls for upload the assets

    2. The client uploads the asset by making a binary `POST` request to the url with the appropriate `Content-Type` header (e.g. image/jpeg).

    3. The client confirms success upload by calling `/documents/{documentId}/assets/finish-upload` endpoint and it returns asset urls for given ids.
host: pages-api-svc
basePath: /v1
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
responses:
  '400':
    description: 'Request was unable to parse, bad JSON or query arguments'
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '403':
    description: Permissions denied
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '404':
    description: Resource not found
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '422':
    description: 'Invalid request, missing required parameters'
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '500':
    description: 'Server encountered a critical error, but the request appears valid'
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
paths:
...
  '/documents/{documentId}/assets/request-upload':
    post:
      tags:
        - Assets
      summary: Creates one or more asset upload requests.
      description: >
        Creates one or more asset upload requests. The response will be an
        array of upload urls that the client needs to upload the asset
        to.
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              assets:
                type: array
                items:
                  $ref: '#/definitions/AssetRequest'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              assets:
                type: array
                items:
                  $ref: '#/definitions/AssetRequestResult'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/documents/{documentId}/assets/finish-upload':
    post:
      tags:
        - Assets
      summary: Mark one or more asset as uploaded
      description: |
        Mark one or more asset as uploaded
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              assetIds:
                type: array
                items:
                  type: string
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              assets:
                type: array
                items:
                  $ref: '#/definitions/Asset'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/documents/{documentId}/assets/{assetId}':
    get:
      tags:
        - Assets
      summary: Gets asset for given `assetId`
      description: |
        Gets asset (id, url, fileName) for given `assetId`
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
        - $ref: '#/parameters/assetId'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Asset'

        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/documents/{documentId}/assets/copy':
    post:
      tags:
        - Assets
      summary: Copy asset from another document and returns new Asset
      description: |
        Copy asset from another document and returns new Asset
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              assetId:
                type: string
                format: uuid
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Asset'

        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/documents/{documentId}/assets/copy-from-url':
    post:
      tags:
        - Assets
      summary: Copy asset url into given document
      description: |
        Copy asset url into given document
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              assetUrl:
                type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Asset'

        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/documents/{documentId}/assets/':
    get:
      tags:
        - Assets
      summary: Gets all assets for given `documentId`
      description: |
        Gets all assets (id, url, fileName) for given `documentId`
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              assets:
                type: array
                items:
                  $ref: '#/definitions/Asset'

        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/documents/{documentId}/assets/external-document/{service}/{serviceAssetId}':
    get:
      tags:
        - Assets
      summary: Gets all assets for given `documentId`
      description: |
        Gets all assets (id, url, fileName) for given `documentId`
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
        - name: service
          description: Service type
          in: path
          type: string
          required: true
        - name: serviceAssetId
          description: Service asset id
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ExternalDocumentResponse'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/documents/{documentId}/assets/flat-prototype':
    get:
      tags:
        - Assets
      summary: Flat prototype by url
      description: |
        Flat prototype by url
      parameters:
        - $ref: '#/parameters/documentId'
        - $ref: '#/parameters/teamIdInQuery'
        - $ref: '#/parameters/userIdInQuery'
        - name: url
          description: Flat prototype url
          in: query
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/PrototypeResponse'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
...
definitions:
...
  AssetRequest:
    type: object
    properties:
      fileName:
        type: string
        description: Asset file name
  AssetRequestResult:
    type: object
    properties:
      id:
        type: string
        format: uuid
        description: The ID of the Asset record
      fileName:
        type: string
        description: Asset file name
      url:
        type: string
        description: Url for upload
  Asset:
    type: object
    properties:
      id:
        type: string
        format: uuid
        description: The ID of the Asset record
      fileName:
        type: string
        description: Asset file name
      url:
        type: string
        description: Asset url
...
  ExternalDocument:
    type: object
    required:
    - name
    - thumbnailUrl
    properties:
      name:
        type: string
      thumbnailUrl:
        type: string
  ExternalDocumentResponse:
    type: object
    required:
    - externalDocument
    properties:
      externalDocument:
        $ref: '#/definitions/ExternalDocument'
  PrototypeResponse:
    type: object
    required:
    - name
    - isMobile
    - width
    - height
    - updatedAt
    properties:
      name:
        type: string
      isMobile:
        type: boolean
      width:
        type: number
      height:
        type: number
      updatedAt:
        type: string
        format: date-time
        example: '2018-02-28T21:39:18.000Z'
      thumbnail:
        type: string
...
parameters:
...
  userIdInQuery:
    name: user_id
    in: query
    description: The ID of the User
    type: integer
    required: true
  teamIdInQuery:
    name: team_id
    in: query
    description: The ID of the Team
    type: string
    required: true
...
  documentId:
    name: documentId
    in: path
    description: The ID of the Document
    required: true
    type: string
...
  assetId:
    name: assetId
    in: path
    description: The ID of the Asset record
    required: true
    type: string
    format: uuid
...